chore: remove RefFreeMetric, SystemSubsetSelector
